+++ src/storage/theme.js
@@ -1,8 +1,4 @@
-const Bus = require("../bus")
 const THEME_NAMES = require("./theme-names")
-const Lens = require("../util/lens")
-const Theme = require("../storage/theme")
-const Session = require("../session")
 
 exports.changeTheme = function (data) {
   const stylesheet = document.querySelector(
