+++ src/vimish/commands/config.js
@@ -1,0 +1,6 @@
+const Command = require("../command")
+const Storage = require("../../storage")
+
+exports.config = Command.of([], async () => {
+  Storage.openInEditor()
+})
