+++ src/vimish/commands/reload-skin.js
@@ -1,0 +1,11 @@
+const Command = require("../command")
+const Bus = require("../../bus")
+const CustomCSS = require("../../storage/custom-css")
+
+exports["reload-skin"] = Command.of([], async () => {
+  await CustomCSS.injectCSS()
+  Bus.emit(Bus.events.FLASH, {
+    kind: "ok",
+    message: "successfully reloaded your skin",
+  })
+})
