+++ src/vimish/commands/ui.js
@@ -1,0 +1,11 @@
+const Command = require("../command")
+const Settings = require("../../settings")
+const Bus = require("../../bus")
+
+exports.ui = Command.of(
+  ["component", "state"],
+  async ({ component, state }) => {
+    Settings.set(`ui.${component}`, state)
+    Bus.emit(Bus.events.REDRAW)
+  }
+)
